# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: API REST para sistema de agendamento de cadeiras de massagem
  title: Sistema de Agendamento API
  version: "1.0"
paths:
  /audit-logs:
    get:
      consumes:
      - application/json
      description: Lista logs de auditoria com paginação e filtros opcionais
      parameters:
      - default: 10
        description: Limite de registros por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por ID do usuário
        in: query
        name: user_id
        type: integer
      - description: Filtrar por ação
        in: query
        name: action
        type: string
      - description: Filtrar por recurso
        in: query
        name: resource
        type: string
      - description: Filtrar por ID do recurso
        in: query
        name: resource_id
        type: integer
      - description: Filtrar por endereço IP
        in: query
        name: ip_address
        type: string
      - description: Data de início (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Data de fim (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de logs de auditoria
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar logs de auditoria
      tags:
      - audit-logs
  /audit-logs/{id}:
    get:
      consumes:
      - application/json
      description: Retorna um log de auditoria específico pelo ID
      parameters:
      - description: ID do log de auditoria
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Log de auditoria encontrado
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Log não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar log de auditoria
      tags:
      - audit-logs
  /audit-logs/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Lista logs de auditoria de um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: user_id
        required: true
        type: integer
      - default: 10
        description: Limite de registros por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de logs do usuário
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Logs de auditoria de usuário
      tags:
      - audit-logs
...
securityDefinitions:
  Bearer:
    description: Digite 'Bearer' seguido de um espaço e o token JWT obtido no login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"